refactor: remove unnecessary file
